Upgrade Flow to the latest version (^0.90.0)
I had to wait for the GraphQL.js upgrade (see: https://github.com/graphql/graphql-js/issues/1586#issuecomment-454593162). I also removed the private fields from the Flow logger because it was broken. I didn't figure out what was wrong (Flow parser error, but it's fine in the playground) so the best strategy to move forward here is to back up a little bit.

kiwicom-source-id: 3db35d947538b4b9fe988f521dfd711997bfe5af