NPM publisher: use our Babel config to transpile JS and Flow
We are now using Babel to transpile both version of our files (JS and Flow) published on NPM. It's much better this way because we use only one system for everything. Moreover, this removes need for `flow-copy-source` and we can do some transformations even on Flow files which is sometimes necessary for a good interoperability with our other external repositories.

kiwicom-source-id: d265feaf7c3feeb180f872fca3cbface8c213f9f